fix: sync teaching examples with OOP1 notes review
- 34/32: NRVO demo prints object address; timing demo keeps the copy
  alive under -O2 via a volatile read
- 64/65: lround cast for period count, throw instead of returning nan
  from an integer function, per-period FRN spread, duration naming
- 52-stl: out-of-bounds generate and end() dereference
- 71-gcd zero guard, 20-simple_calc divide-by-zero guard,
  50 int/size_t read mismatch, 73 damaged output strings,
  37/40/51/54/62/10 small correctness and sync fixes

+++ 50-binaryinoutfile/main.cpp
@@ -58,7 +58,7 @@ void test_binary() {
     cout << str << '\n';
 
     rf.read(reinterpret_cast<char *>(&size), sizeof(size_t)); // 10
-    vector<size_t> vec(size); // 10 element long vector
+    vector<int> vec(size); // 10 element long vector
     for (auto &v : vec) {     // auto with reference type
       rf.read(reinterpret_cast<char *>(&v),
               sizeof(int)); // read each element into it
